requirements.txt
Flask==2.3.3
gunicorn==25.1.0
google-generativeai==0.8.5
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.2
scipy==1.15.3
matplotlib==3.8.4
pillow==10.4.0
python-dotenv==1.0.1
